Move validators to route package
