+++ pkg/route/config.go
@@ -3,7 +3,6 @@ package route
 import (
 	"github.com/cloudretic/matcha/pkg/cors"
 	"github.com/cloudretic/matcha/pkg/middleware"
-	"github.com/cloudretic/matcha/pkg/validator"
 )
 
 // RouteConfigFuncs can be applied to a Route at creation.
@@ -25,7 +24,7 @@ func WithMiddleware(mw middleware.Middleware) ConfigFunc {
 	}
 }
 
-func Validators(vs ...validator.Validator) ConfigFunc {
+func Validators(vs ...Validator) ConfigFunc {
 	return func(r Route) error {
 		for _, v := range vs {
 			r.Validate(v)
