+++ pkg/route/default.go
@@ -10,7 +10,6 @@ import (
 	"github.com/cloudretic/matcha/pkg/middleware"
 	"github.com/cloudretic/matcha/pkg/path"
 	"github.com/cloudretic/matcha/pkg/rctx"
-	"github.com/cloudretic/matcha/pkg/validator"
 )
 
 //=====PARTS=====
@@ -144,7 +143,7 @@ type defaultRoute struct {
 	method     string
 	parts      []Part
 	middleware []middleware.Middleware
-	validators []validator.Validator
+	validators []Validator
 }
 
 // Tokenize and parse a route expression into a defaultRoute.
@@ -156,7 +155,7 @@ func build_defaultRoute(method, expr string) (*defaultRoute, error) {
 		method:     method,
 		parts:      make([]Part, 0),
 		middleware: make([]middleware.Middleware, 0),
-		validators: make([]validator.Validator, 0),
+		validators: make([]Validator, 0),
 	}
 	var token string
 	for next := 0; next < len(expr); {
@@ -248,7 +247,7 @@ func (route *defaultRoute) Attach(mw middleware.Middleware) {
 	route.middleware = append(route.middleware, mw)
 }
 
-func (route *defaultRoute) Validate(v validator.Validator) {
+func (route *defaultRoute) Validate(v Validator) {
 	route.validators = append(route.validators, v)
 }
 
@@ -256,6 +255,6 @@ func (route *defaultRoute) Middleware() []middleware.Middleware {
 	return route.middleware
 }
 
-func (route *defaultRoute) Validators() []validator.Validator {
+func (route *defaultRoute) Validators() []Validator {
 	return route.validators
 }
