+++ pkg/route/partial.go
@@ -8,7 +8,6 @@ import (
 	"github.com/cloudretic/matcha/pkg/middleware"
 	"github.com/cloudretic/matcha/pkg/path"
 	"github.com/cloudretic/matcha/pkg/rctx"
-	"github.com/cloudretic/matcha/pkg/validator"
 )
 
 // =====PARTS=====
@@ -100,7 +99,7 @@ type partialRoute struct {
 	method     string
 	parts      []Part
 	middleware []middleware.Middleware
-	validators []validator.Validator
+	validators []Validator
 }
 
 // Tokenize and parse a route expression into a partialRoute.
@@ -221,7 +220,7 @@ func (route *partialRoute) Attach(m middleware.Middleware) {
 	route.middleware = append(route.middleware, m)
 }
 
-func (route *partialRoute) Validate(v validator.Validator) {
+func (route *partialRoute) Validate(v Validator) {
 	route.validators = append(route.validators, v)
 }
 
@@ -229,6 +228,6 @@ func (route *partialRoute) Middleware() []middleware.Middleware {
 	return route.middleware
 }
 
-func (route *partialRoute) Validators() []validator.Validator {
+func (route *partialRoute) Validators() []Validator {
 	return route.validators
 }
