+++ pkg/route/route.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 
 	"github.com/cloudretic/matcha/pkg/middleware"
-	"github.com/cloudretic/matcha/pkg/validator"
 )
 
 type Route interface {
@@ -42,11 +41,11 @@ type Route interface {
 	// Attach a validator to the route.
 	//
 	// Validators cannot be removed from a router once they are added.
-	Validate(v validator.Validator)
+	Validate(v Validator)
 	// Get the middleware attached to the route.
 	Middleware() []middleware.Middleware
 	// Get the validators attached to the route.
-	Validators() []validator.Validator
+	Validators() []Validator
 }
 
 // Create a new Route based on a string expression.
