+++ pkg/router/router_test.go
@@ -13,7 +13,6 @@ import (
 	"sync"
 
 	"github.com/cloudretic/matcha/pkg/cors"
-	"github.com/cloudretic/matcha/pkg/validator"
 
 	"github.com/cloudretic/matcha/pkg/rctx"
 	"github.com/cloudretic/matcha/pkg/route"
@@ -357,7 +356,7 @@ func TestValidatedDuplicate(t *testing.T) {
 		Default(),
 		WithRoute(route.Declare(
 			http.MethodGet, "/",
-			route.Validators(validator.Hosts("origin.com")),
+			route.Validators(route.Hosts("origin.com")),
 		), h1),
 		WithRoute(route.Declare(http.MethodGet, "/"), h2),
 	)
