+++ pkg/tree/tree.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/cloudretic/matcha/pkg/path"
 	"github.com/cloudretic/matcha/pkg/route"
-	"github.com/cloudretic/matcha/pkg/validator"
 )
 
 const NO_LEAF_ID = int(0)
@@ -17,7 +16,7 @@ type node struct {
 	p               route.Part
 	children        []*node
 	leaf_id         int
-	leaf_validators []validator.Validator
+	leaf_validators []route.Validator
 }
 
 func (n *node) isLeaf() bool {
@@ -35,7 +34,7 @@ func (n *node) resolveLeafForRequest(req *http.Request) int {
 	if n.leaf_id == NO_LEAF_ID {
 		return NO_LEAF_ID
 	}
-	if !validator.ExecuteValidators(req, n.leaf_validators) {
+	if !route.ExecuteValidators(req, n.leaf_validators) {
 		return NO_LEAF_ID
 	}
 	return n.leaf_id
