--- pkg/validator/validator_test.go
@@ -1,9 +1,0 @@
-package validator
-
-import (
-	"testing"
-)
-
-func TestHostValidator(t *testing.T) {
-
-}
